// src/types/dashboard-summary.d.ts
export interface DashboardSummary {
    totalRegisteredUsers: number;
    totalActiveAccounts: number;
    recentTransactions: number;
    pendingApproval: number;
    percentageIncreaseInTotalRegisteredUsers: number;
    percentageIncreaseInTotalActiveAccounts: number;
    percentageIncreaseInRecentTransactions: number;
    percentageIncreaseInPendingApproval: number;
}